+++ include/xgboost/generic_parameters.h
@@ -9,17 +9,12 @@
 #include <xgboost/enum_class_param.h>
 
 namespace xgboost {
-enum class TreeMethod : int {
-  kAuto = 0, kApprox = 1, kExact = 2, kHist = 3,
-  kGPUExact = 4, kGPUHist = 5
-};
 
 enum class DataSplitMode : int {
   kAuto = 0, kCol = 1, kRow = 2
 };
 }  // namespace xgboost
 
-DECLARE_FIELD_ENUM_CLASS(xgboost::TreeMethod);
 DECLARE_FIELD_ENUM_CLASS(xgboost::DataSplitMode);
 
 namespace xgboost {
@@ -30,8 +25,6 @@ struct LearnerTrainParam : public dmlc::Parameter<LearnerTrainParam> {
   bool seed_per_iteration;
   // data split mode, can be row, col, or none.
   DataSplitMode dsplit;
-  // tree construction method
-  TreeMethod tree_method;
   // number of threads to use if OpenMP is enabled
   // if equals 0, use system default
   int nthread;
@@ -41,6 +34,8 @@ struct LearnerTrainParam : public dmlc::Parameter<LearnerTrainParam> {
   int gpu_id;
   // number of devices to use, -1 implies using all available devices.
   int n_gpus;
+
+  std::string booster;
 
   // declare parameters
   DMLC_DECLARE_PARAMETER(LearnerTrainParam) {
@@ -58,15 +53,6 @@ struct LearnerTrainParam : public dmlc::Parameter<LearnerTrainParam> {
         .add_enum("col", DataSplitMode::kCol)
         .add_enum("row", DataSplitMode::kRow)
         .describe("Data split mode for distributed training.");
-    DMLC_DECLARE_FIELD(tree_method)
-        .set_default(TreeMethod::kAuto)
-        .add_enum("auto", TreeMethod::kAuto)
-        .add_enum("approx", TreeMethod::kApprox)
-        .add_enum("exact", TreeMethod::kExact)
-        .add_enum("hist", TreeMethod::kHist)
-        .add_enum("gpu_exact", TreeMethod::kGPUExact)
-        .add_enum("gpu_hist", TreeMethod::kGPUHist)
-        .describe("Choice of tree construction method.");
     DMLC_DECLARE_FIELD(nthread).set_default(0).describe(
         "Number of threads to use.");
     DMLC_DECLARE_FIELD(disable_default_eval_metric)
@@ -79,6 +65,9 @@ struct LearnerTrainParam : public dmlc::Parameter<LearnerTrainParam> {
         .set_default(0)
         .set_lower_bound(-1)
         .describe("Number of GPUs to use for multi-gpu algorithms.");
+    DMLC_DECLARE_FIELD(booster)
+        .set_default("gbtree")
+        .describe("Gradient booster used for training.");
   }
 };
 }  // namespace xgboost
